+++ multi_region_tests.py
@@ -6,6 +6,7 @@ import dxpy
 import json
 import pprint
 import os
+import time
 import sys
 import subprocess
 import util
@@ -21,24 +22,53 @@ top_dir = os.path.dirname(sys.argv[0])
 test_dir = os.path.join(os.path.abspath(top_dir), "test")
 target_folder = "/release_test"
 
+def wait_for_completion(test_exec_objs):
+    print("awaiting completion ...")
+    # wait for analysis to finish while working around Travis 10m console inactivity timeout
+    noise = subprocess.Popen(["/bin/bash", "-c", "while true; do sleep 60; date; done"])
+    try:
+        for exec_obj in test_exec_objs:
+            exec_obj.wait_on_done()
+    finally:
+        noise.kill()
+    print("done")
+
+# Run [workflow] on several inputs, return the analysis ID.
+def run_workflow(dx_proj, test_folder, oid):
+    def once():
+        try:
+            dx_proj.new_folder(test_folder, parents=True)
+            wf = dxpy.DXWorkflow(project=dx_proj.get_id(), dxid=oid)
+            return wf.run({},
+                          project=dx_proj.get_id(),
+                          folder=test_folder)
+        except Exception, e:
+            print("exception message={}".format(e))
+            return None
+
+    for i in range(1,5):
+        retval = once()
+        if retval is not None:
+            return retval
+        print("Sleeping for 5 seconds before trying again")
+        time.sleep(5)
+    raise RuntimeError("running workflow")
+
 # Build a workflow.
 #
 # wf             workflow name
 # classpath      java classpath needed for running compilation
 # folder         destination folder on the platform
-def build_test(source_file, proj_name, folder, version_id):
-    dx_proj = util.get_project(proj_name)
-    if dx_proj is None:
-        raise RuntimeError("Could not find project {}".format(proj_name))
+def build_test(source_file, dx_proj, folder, version_id):
     dx_proj.new_folder(folder, parents=True)
-    print("Compiling {} to project {}:/{}".format(source_file, proj_name, folder))
+    print("Compiling {} to project {}:/{}".format(source_file, dx_proj.name, folder))
     cmdline = [ "java", "-jar",
                 os.path.join(top_dir, "dxWDL-{}.jar".format(version_id)),
                 "compile",
                 source_file,
                 "-force",
                 "-folder", folder,
-                "-project", proj_name ]
+                "-project", dx_proj.get_id() ]
     print(" ".join(cmdline))
     oid = subprocess.check_output(cmdline).strip()
     return oid
@@ -50,11 +80,21 @@ def main():
     version_id = util.get_version_id(top_dir)
     wdl_source_file = os.path.join(test_dir, "basic/trivial.wdl")
     dx_objects = []
-    for proj_name in projects:
-        oid = build_test(wdl_source_file, proj_name, target_folder, version_id)
-        dx_objects.append(oid)
+    test_exec_objs=[]
 
-    #oid = build_test(tname, project, applet_folder, version_id, c_flags)
+    # build version of the applet on all regions
+    for proj_name in projects:
+        dx_proj = util.get_project(proj_name)
+        if dx_proj is None:
+            raise RuntimeError("Could not find project {}".format(proj_name))
+        oid = build_test(wdl_source_file, dx_proj, target_folder, version_id)
+        anl = run_workflow(dx_proj, target_folder, oid)
+        print("Running {}".format(oid))
+        test_exec_objs.append(anl)
+
+    # Wait for completion
+    print("executables: " + ", ".join([a.get_id() for a in test_exec_objs]))
+    wait_for_completion(test_exec_objs)
 
 if __name__ == '__main__':
     main()