+++ src/main/scala/dxWDL/compiler/GenerateIR.scala
@@ -893,7 +893,7 @@ task Add {
 
         // Create a stage per call/scatter-block/declaration-block
         val van = new VarAnalysis(Set.empty, Map.empty, cef, verbose)
-        val subBlocks = Block.splitIntoBlocks(wfProper.toVector, locked, van)
+        val subBlocks = Block.splitIntoBlocks(wfProper.toVector, van)
 
         val (allStageInfo_i, wfOutputs) =
             if (locked)